ProtectingProjects

+++ api/src/controllers/project/04 - deleteProject.ts
@@ -1,13 +1,15 @@
 import Project from "../../models/Project";
 import { idSchema } from "../../schema/project";
 
-export const deleteProject = async (id:string) => {
+export const deleteProject = async (id:string,userId?:string) => {
   const result = idSchema.safeParse(id);
   if(!result.success) throw new Error(`ID invalid`);
 
   const projectDeleted = await Project.findById(id);
 
   if(!projectDeleted) throw new Error(`Project not found`)
+  if( userId && projectDeleted.manager.toString() !== userId.toString()) throw new Error('Unauthorized');
+
   await projectDeleted.deleteOne();
 
   return {
